scraper/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
